[src]:branches/VS/obsproc_prep/sorc
M    load_libs.rc

Updating loab_libs.rc to set w3emc_ver to v2.2.0 (replacing v2.0.5) in anticipation of the soon to be implemented update to the w3emc library that was released to NCO May 1, 2014 and will be implemented prior to obsproc phase 2 release.

M    setlibs.rc

Correcting a typo in changeset r40069 re: w3emc_ver.  It is now set to v2.2.0, replacing the v2.0.6 typo.