[src]:branches/VS/obsproc_prep
Reintegration merge of the development branch:
  `branches/obsproc_prep/obsproc_prep.tkt-352.crayport`
into its trunk, `branches/VS/obsproc_prep`.

This update allows obsproc_prep software to be compiled and run on the WCOSS
Cray-XC40 as well as WCOSS IBM iDataPlex.

A couple of bug fixes are included along with an increase in the max allowable
number of stations to be qc'd in prepobs_cqcvad and prepobs_profcqc.

Also, the variable name specifying optional launching mechanism for prepbufr 
processing was changed from "launcher" to the more specific "launcher_PREP".
(This change has no impact on currrent production obsproc prep jobs).


For more details, see ticket #352.